backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9

sentence-transformers==3.0.1
torch>=2.2.0

chromadb==0.5.5

PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5

python-dotenv==1.0.1
transformers
accelerate
bitsandbytes
